+++ src/zigSetup.ts
@@ -172,13 +172,10 @@ export async function setupZig(context: ExtensionContext) {
     });
 
     const configuration = workspace.getConfiguration("zig", null);
-    if (!configuration.get<string | null>("zigPath", null)) {
-        const zigInPath = which.sync("zig", { nothrow: true });
-        const options = ["Install", "Specify path"];
-        if (zigInPath) options.push("Use Zig in PATH");
+    if (configuration.get<string | null>("zigPath", null) === null) {
         const response = await window.showInformationMessage(
             "Zig path hasn't been set, do you want to specify the path or install Zig?",
-            ...options
+            "Install", "Specify path", "Use Zig in PATH"
         );
 
         if (response === "Install") {
@@ -200,7 +197,7 @@ export async function setupZig(context: ExtensionContext) {
                 await configuration.update("zigPath", uris[0].fsPath, true);
             }
         } else if (response == "Use Zig in PATH") {
-            await configuration.update("zigPath", zigInPath, true);
+            await configuration.update("zigPath", "", true);
         } else throw "zigPath not specified";
     }